Edit Employees

+++ controllers/employeeController.js
@@ -1,8 +1,8 @@
 const mongoose = require('mongoose');
 const AsyncHandler = require('express-async-handler');
-const ApiError = require('../utils/apiError');
 const { dirname } = require('path');
 const fs = require('fs');
+const ApiError = require('../utils/apiError');
 require('../models/Employee');
 
 const Employees = mongoose.model('employees');
@@ -14,7 +14,7 @@ exports.getAllEmployees = AsyncHandler( async (req, res, next) => {
     const allEmployees = await Employees.find({});
 
     if(!allEmployees){
-        return new ApiError('no Employee found', 404);
+        throw new ApiError('no Employee found', 404);
     }
 
     res.status(200).json({data: allEmployees});
@@ -25,7 +25,7 @@ exports.getEmployeeById = AsyncHandler( async (req, res, next) => {
     const employee = await Employees.findById(req.params.id);
 
     if(!employee){
-        return new ApiError('no Employee found', 404);
+        throw new ApiError('no Employee found', 404);
     }
 
     res.status(200).json({data: employee});
@@ -50,7 +50,7 @@ exports.createEmployee = AsyncHandler( async (req, res, next) => {
     const employee = await Employees.create(obj);
 
     if(!employee){
-        return new ApiError('Error happened while Creating Employee', 404);
+        throw new ApiError('Error happened while Creating Employee', 404);
     }
 
     res.status(200).json({data: employee});
@@ -59,9 +59,8 @@ exports.createEmployee = AsyncHandler( async (req, res, next) => {
 
 // Update Employee
 exports.updateEmployee = AsyncHandler( async (req, res, next) => {
-    if(req.file !== undefined){
-        req.body.image = req.file.filename;
-    }
+    if (req.file !== undefined) req.body.image = req.file.filename;
+
     const employee = await Employees.findOneAndUpdate({_id: req.params.id},{
         $set: {
             name: req.body.name,
@@ -78,16 +77,18 @@ exports.updateEmployee = AsyncHandler( async (req, res, next) => {
     });
 
     if(!employee){
-        const root = dirname(require.main.filename);
-        const path = root + "/images/Employee/" + req.body.image;
-        fs.unlink(path, (err) => {
-            if (err) {
-                console.log(err);
-            }
-        });
-        return new ApiError('Error happened while Updating Employee', 404);
+        if (req.file !== undefined){
+            const root = dirname(require.main.filename);
+            const path = root + "/images/Employee/" + req.body.image;
+            fs.unlink(path, (err) => {
+                if (err) {
+                    console.log(err);
+                }
+            });
+        }
+        console.log(employee);
+        throw new ApiError('Error happened while Updating Employee', 404);
     }
-
     if(req.file !== undefined){
         const root = dirname(require.main.filename);
         const path = root + "/images/Employee/" + employee.image;
@@ -99,6 +100,7 @@ exports.updateEmployee = AsyncHandler( async (req, res, next) => {
     }
 
     res.status(200).json({mssg:'Updated',oldData: employee});
+
 });
 
 
@@ -107,7 +109,7 @@ exports.deleteEmployee = AsyncHandler( async (req, res, next) => {
     const employee = await Employees.findOneAndDelete({_id: req.params.id});
 
     if(!employee){
-        return new ApiError('Error happened while Deleting Employee', 404);
+        throw new ApiError('Error happened while Deleting Employee', 404);
     }
 
     const root = dirname(require.main.filename);
@@ -131,7 +133,7 @@ exports.resetPassword = AsyncHandler( async (req, res, next) => {
     });
 
     if(!employee){
-        return new ApiError('Error happened while Resetting Password', 404);
+        throw new ApiError('Error happened while Resetting Password', 404);
     }
 
     res.status(200).json({mssg:'Password Reseted',oldData: employee});
@@ -148,7 +150,7 @@ exports.banEmployee = AsyncHandler( async (req, res, next) => {
     });
 
     if(!employee){
-        return new ApiError('Error happened while Banning Employee', 404);
+        throw new ApiError('Error happened while Banning Employee', 404);
     }
 
     res.status(200).json({mssg:'Employee Banned'});
@@ -163,7 +165,7 @@ exports.unbanEmployee = AsyncHandler( async (req, res, next) => {
     });
 
     if(!employee){
-        return new ApiError('Error happened while Unbanning Employee', 404);
+        throw new ApiError('Error happened while Unbanning Employee', 404);
     }
 
     res.status(200).json({mssg:'Employee Unbanned'});
@@ -178,7 +180,7 @@ exports.deactivateEmployee = AsyncHandler( async (req, res, next) => {
     });
 
     if(!employee){
-        return new ApiError('Error happened while Deactivating Employee', 404);
+        throw new ApiError('Error happened while Deactivating Employee', 404);
     }
 
     res.status(200).json({mssg:'Employee Deactivated'});
@@ -193,8 +195,21 @@ exports.activateEmployee = AsyncHandler( async (req, res, next) => {
     });
 
     if(!employee){
-        return new ApiError('Error happened while Activating Employee', 404);
+        throw new ApiError('Error happened while Activating Employee', 404);
     }
 
     res.status(200).json({mssg:'Employee Activated'});
+});
+
+// filter Employee
+exports.filterEmployee = AsyncHandler( async (req, res, next) => {
+    const queryStr = JSON.stringify(req.query);
+
+    const employees = await Employees.find(JSON.parse(queryStr));
+
+    if(!employees){
+        throw new ApiError('Error happened while Filtering Employee', 404);
+    }
+
+    res.status(200).json({data: employees});
 });
